Logging stack traces on failed imports
This should really use Monolog, so that we can send the logs also to files and format them appropriately, etc.

This is the minimal change to be able to debug what is happening in the end2end environment.

Unfortunately there seem to be no generic way of dumping the identifier for a Model object from the SDK, so I can't put a `getId()` in the message as some Models do not have it.